[TASK] Update coding styleguide to PER + additional configs
This change adapts our coding guidelines
and rules with PHP-CS-Fixer, allowing for
* PER (a.k.a. PSR-12)
* no_useless_nullsafe_operator (only use nullsafe if allowed)
* only single-spaces in arrays

This effectively adapts our code base
* use "public const" instead of "const" where visibility is not defined
* one whitespace between "declare()" and "namespace" declarations
* Use (float) instead of (double) casting
* Additional space in annonymous classes (PER)
* grouped "use" statements (classes, function, constants)
* only one space between items in array lists
* Use (bool) instead of (boolean)

Resolves: #98585
Releases: main, 11.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76097
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/backend/Classes/Form/Element/NumberElement.php
@@ -174,7 +174,7 @@ class NumberElement extends AbstractFormElement
         $valueSliderHtml = [];
         if (is_array($config['slider'] ?? false)) {
             if ($format === 'decimal') {
-                $itemValue = (double)$itemValue;
+                $itemValue = (float)$itemValue;
             } else {
                 $itemValue = (int)$itemValue;
             }
